Move Configuration from common.cost to common
