+++ app/src/main/java/org/astraea/app/scenario/ScenarioMain.java
@@ -19,10 +19,10 @@ package org.astraea.app.scenario;
 import com.beust.jcommander.Parameter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.Admin;
 import org.astraea.common.argument.Argument;
-import org.astraea.common.cost.Configuration;
 import org.astraea.common.scenario.Scenario;
 
 public class ScenarioMain extends Argument {
