+++ common/src/main/java/org/astraea/common/Configuration.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.cost;
+package org.astraea.common;
 
 import java.util.Arrays;
 import java.util.Collections;
