+++ common/src/main/java/org/astraea/common/Utils.java
@@ -30,7 +30,6 @@ import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import org.astraea.common.cost.Configuration;
 
 public final class Utils {
