+++ common/src/main/java/org/astraea/common/balancer/algorithms/AlgorithmConfig.java
@@ -25,10 +25,10 @@ import java.util.UUID;
 import java.util.function.BiPredicate;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
+import org.astraea.common.Configuration;
 import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.balancer.Balancer;
 import org.astraea.common.cost.ClusterCost;
-import org.astraea.common.cost.Configuration;
 import org.astraea.common.cost.HasClusterCost;
 import org.astraea.common.cost.HasMoveCost;
 import org.astraea.common.cost.MoveCost;
