+++ common/src/main/java/org/astraea/common/partitioner/Dispatcher.java
@@ -22,10 +22,10 @@ import java.util.stream.Collectors;
 import org.apache.kafka.clients.producer.Partitioner;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.common.Cluster;
+import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.ReplicaInfo;
-import org.astraea.common.cost.Configuration;
 
 public interface Dispatcher extends Partitioner {
   /**
