+++ common/src/main/java/org/astraea/common/partitioner/StrictCostDispatcher.java
@@ -25,13 +25,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.ReplicaInfo;
 import org.astraea.common.argument.DurationField;
 import org.astraea.common.cost.BrokerCost;
-import org.astraea.common.cost.Configuration;
 import org.astraea.common.cost.HasBrokerCost;
 import org.astraea.common.cost.NodeLatencyCost;
 import org.astraea.common.metrics.collector.MetricCollector;
