+++ common/src/test/java/org/astraea/common/balancer/algorithms/GreedyBalancerTest.java
@@ -20,10 +20,10 @@ import java.time.Duration;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.IntStream;
+import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.balancer.Balancer;
 import org.astraea.common.balancer.FakeClusterInfo;
-import org.astraea.common.cost.Configuration;
 import org.astraea.common.cost.DecreasingCost;
 import org.astraea.common.metrics.BeanQuery;
 import org.astraea.common.metrics.MBeanClient;
