+++ common/src/test/java/org/astraea/common/cost/DecreasingCost.java
@@ -16,6 +16,7 @@
  */
 package org.astraea.common.cost;
 
+import org.astraea.common.Configuration;
 import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.Replica;
